refactor(note): 收敛笔记工作区控制逻辑
