+++ src/views/workspace/note/_components/NoteWorkspace/index.tsx
@@ -1,47 +1,25 @@
 ﻿import { Spin } from '@/components/Feedback';
 import InlineComment from '@/components/InlineComment';
 import SegmentedTabs from '@/components/SegmentedTabs';
-import { useMemoizedFn, useRequest, useUnmount } from 'ahooks';
-import { useEffect, useRef, useState, useSyncExternalStore } from 'react';
+import { useMemoizedFn, useUnmount } from 'ahooks';
+import { useEffect, useRef, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 
 import CustomBlockNote from '@/components/Note/CustomBlockNote';
 import type {
   NoteBodyEditorHandle,
-  NoteCollaborationUser,
   NoteOutlineItem,
 } from '@/components/Note/CustomBlockNote/index.type';
-import { useInlineCommentService, useInteractService, useUserService } from '@/domains';
-import type {
-  AiDiffDisplayMode,
-  NoteInfoDisplayData,
-  NoteInlineCommentDraft,
-  NoteSaveStatus,
-  NoteSelectionSnapshot,
-} from '@/domains/Note';
-import {
-  AI_DIFF_DISPLAY_MODE,
-  encodeNoteClientContentSignature,
-  NoteInlineCommentSession,
-  useNoteSession,
-} from '@/domains/Note';
-import type { User } from '@/domains/User';
+import type { AiDiffDisplayMode, NoteInfoDisplayData } from '@/domains/Note';
+import { AI_DIFF_DISPLAY_MODE, encodeNoteClientContentSignature } from '@/domains/Note';
 import { useResourceDisplayName } from '@/hooks/useResourceDisplayName';
-import { useSmoothFlag } from '@/hooks/useSmoothFlag';
-import { parseErrorMessage } from '@/utils/error';
 import { RESOURCE_KIND } from '@/utils/navigation/resourceTarget';
 import {
-  useResourceHostChatContextActions,
   useResourceHostLayoutConfig,
   type ResourceHostLayoutConfig,
 } from '@/views/workspace/ResourceHostContext';
-import { useWorkspaceResourceSidePanelStore } from '@/views/workspace/_store/useWorkspaceResourceSidePanelStore';
-import { Alert, Button, toast } from '@heroui/react';
+import { Alert, Button } from '@heroui/react';
 import { History } from 'lucide-react';
-import {
-  createNoteChatStateProvider,
-  createNoteSelectionChatContext,
-} from '../../NoteChatProtocol';
 import { useNoteFindMode } from '../../_hooks/useNoteFindMode';
 import { useAiDiffDisplayStore } from '../../_store/useAiDiffDisplayStore';
 import styles from '../../style.module.less';
@@ -49,6 +27,8 @@ import FindBar from '../FindBar';
 import NoteInfoBar from '../NoteInfoBar';
 import NoteOutline, { NOTE_OUTLINE_TITLE_ID } from '../NoteOutline';
 import NoteTitle, { type NoteTitleHandle, type NoteTitleSaveStatus } from '../NoteTitle';
+import { resolveNoteHeaderSaveStatus } from './noteWorkspaceModel';
+import { useNoteWorkspaceController } from './useNoteWorkspaceController';
 
 interface NoteWorkspaceProps {
   resourceId: string;
@@ -56,88 +36,10 @@ interface NoteWorkspaceProps {
   onRefreshNoteInfo: () => unknown | Promise<unknown>;
 }
 
-const INLINE_COMMENT_POLLING_INTERVAL = 8_000;
-
-const NOTE_COLLABORATION_COLORS = [
-  '#2563eb',
-  '#16a34a',
-  '#dc2626',
-  '#d97706',
-  '#7c3aed',
-  '#0891b2',
-  '#db2777',
-  '#4f46e5',
-  '#059669',
-  '#ea580c',
-] as const;
-
-function getNoteCollaborationUserName(user: User | undefined, fallbackName: string): string {
-  return user?.nickname?.trim() || user?.realName?.trim() || user?.username?.trim() || fallbackName;
-}
-
-function pickNoteCollaborationColor(seed: string): string {
-  let hash = 0;
-  for (let i = 0; i < seed.length; i += 1) {
-    hash = (hash * 31 + seed.charCodeAt(i)) >>> 0;
-  }
-  return NOTE_COLLABORATION_COLORS[hash % NOTE_COLLABORATION_COLORS.length];
-}
-
-function sanitizeDownloadFileName(fileName: string, fallbackName: string): string {
-  const normalizedName = fileName.trim().replace(/[\\/:*?"<>|]+/g, '_');
-  const safeName = normalizedName.replace(/[.\s]+$/g, '');
-  return safeName || fallbackName;
-}
-
-function downloadTextArtifact(params: {
-  content: string;
-  mimeType: string;
-  fileName: string;
-}): void {
-  const blob = new Blob([params.content], { type: params.mimeType });
-  const url = URL.createObjectURL(blob);
-  const anchor = document.createElement('a');
-  anchor.href = url;
-  anchor.download = params.fileName;
-  anchor.style.display = 'none';
-  document.body.appendChild(anchor);
-  try {
-    anchor.click();
-  } finally {
-    document.body.removeChild(anchor);
-    URL.revokeObjectURL(url);
-  }
-}
-
-function buildNoteCollaborationUser(
-  user: User | undefined,
-  fallbackName: string
-): NoteCollaborationUser {
-  const name = getNoteCollaborationUserName(user, fallbackName);
-  const colorSeed = user?.id?.trim() || user?.username?.trim() || name;
-  return {
-    name,
-    color: pickNoteCollaborationColor(colorSeed),
-  };
-}
-
-type NoteHeaderSaveStatus = NoteSaveStatus | 'failed';
-
-function resolveNoteHeaderSaveStatus(
-  bodyStatus: NoteSaveStatus,
-  titleStatus: NoteTitleSaveStatus
-): NoteHeaderSaveStatus {
-  if (titleStatus === 'failed') return 'failed';
-  if (bodyStatus === 'waiting') return 'waiting';
-  if (bodyStatus === 'saving' || titleStatus === 'saving') return 'saving';
-  return 'saved';
-}
-
 function NoteWorkspace({ resourceId, noteInfoDisplay, onRefreshNoteInfo }: NoteWorkspaceProps) {
   const { t } = useTranslation('note');
   const aiDiffDisplayMode = useAiDiffDisplayStore((state) => state.displayMode);
   const setAiDiffDisplayMode = useAiDiffDisplayStore((state) => state.setDisplayMode);
-  const { setChatContext } = useResourceHostChatContextActions();
   const bodyEditorRef = useRef<NoteBodyEditorHandle>(null);
   const titleEditorRef = useRef<NoteTitleHandle>(null);
   const mainScrollRef = useRef<HTMLDivElement>(null);
@@ -150,43 +52,8 @@ function NoteWorkspace({ resourceId, noteInfoDisplay, onRefreshNoteInfo }: NoteW
   const [activeHeadingId, setActiveHeadingId] = useState<string | undefined>(undefined);
   const [aiBulkActionsPortalContainer, setAiBulkActionsPortalContainer] =
     useState<HTMLDivElement | null>(null);
-  const [exportPending, setExportPending] = useState(false);
   const [titleSaveStatus, setTitleSaveStatus] = useState<NoteTitleSaveStatus>('saved');
   const [hasAiDiffContent, setHasAiDiffContent] = useState(false);
-  const [inlineCommentDraft, setInlineCommentDraft] = useState<NoteInlineCommentDraft>();
-  const [activeInlineCommentThreadId, setActiveInlineCommentThreadId] = useState<string>();
-  const [inlineCommentScrollTarget, setInlineCommentScrollTarget] = useState<{
-    threadId: string;
-  }>();
-  const [isInlineCommentHistoryOpen, setIsInlineCommentHistoryOpen] = useState(false);
-  const interactService = useInteractService();
-  const inlineCommentService = useInlineCommentService();
-  const userService = useUserService();
-  const setResourceSidePanelMode = useWorkspaceResourceSidePanelStore((state) => state.setMode);
-  const [inlineCommentSession] = useState(
-    () => new NoteInlineCommentSession({ resourceId, inlineCommentService })
-  );
-  const inlineCommentSnapshot = useSyncExternalStore(
-    inlineCommentSession.subscribe,
-    inlineCommentSession.getSnapshot
-  );
-  const { data: currentUser, error: currentUserError } = useRequest(() =>
-    userService.getUserInfo()
-  );
-  const shouldWaitCurrentUser = !currentUser && !currentUserError;
-  const { status, saveStatus, doc, provider, reconnect, idbSynced } = useNoteSession(resourceId, {
-    actorUserId: currentUser?.id,
-    enabled: !shouldWaitCurrentUser,
-    localOnly: Boolean(noteInfoDisplay.aiDiffPreview),
-  });
-  const isConnected = status === 'connected';
-  const isDisconnected = useSmoothFlag(status === 'disconnected', 2000, 2000);
-  const isEditorReadOnly = status === 'connecting' || !noteInfoDisplay.canCollaborativeEdit;
-  const isTitleReadOnly = !noteInfoDisplay.canCollaborativeEdit;
-  const blockLocalDocWrites = isConnected && !noteInfoDisplay.canCollaborativeEdit;
-  const showFullPageSpin = (status === 'connecting' && !idbSynced) || shouldWaitCurrentUser;
-  const middleOverlayText =
-    status === 'connecting' && !idbSynced ? t('workspace.connecting') : t('workspace.loadingUser');
   const fallbackNoteTitle = noteInfoDisplay.noteTitle;
   const [aiDiffBodyContentHash, setAiDiffBodyContentHash] = useState<string | undefined>(undefined);
   const noteClientContentSignature = aiDiffBodyContentHash
@@ -195,26 +62,57 @@ function NoteWorkspace({ resourceId, noteInfoDisplay, onRefreshNoteInfo }: NoteW
   const isNoteClientContentSignaturePending = !aiDiffBodyContentHash;
   const untitledTitle = t('title.untitled');
   const resourceName = useResourceDisplayName(resourceId, fallbackNoteTitle, untitledTitle);
+  const workspace = useNoteWorkspaceController({
+    bodyEditorRef,
+    fallbackNoteTitle,
+    isNoteClientContentSignaturePending,
+    noteClientContentSignature,
+    noteInfoDisplay,
+    resourceId,
+    t,
+    titleEditorRef,
+    untitledTitle,
+  });
+  const {
+    activeInlineCommentThreadId,
+    blockLocalDocWrites,
+    canRenderBodyEditor,
+    collaborationUser,
+    currentUser,
+    doc,
+    exportPending,
+    handleAskAi,
+    handleDownloadMarkdown,
+    handleInlineCommentThreadSelect,
+    handlePrintPdf,
+    inlineCommentDraft,
+    inlineCommentScrollTarget,
+    inlineCommentSession,
+    inlineCommentSnapshot,
+    inlineCommentsBinding,
+    isConnected,
+    isDisconnected,
+    isEditorReadOnly,
+    isInlineCommentHistoryOpen,
+    isTitleReadOnly,
+    middleOverlayText,
+    noteChatStateProvider,
+    provider,
+    reconnect,
+    saveStatus,
+    setActiveInlineCommentThreadId,
+    setInlineCommentDraft,
+    setIsInlineCommentHistoryOpen,
+    showFullPageSpin,
+    status,
+  } = workspace;
   const headerSaveStatus = resolveNoteHeaderSaveStatus(saveStatus, titleSaveStatus);
   const saveStatusText = t(`save.${headerSaveStatus}`);
-  const collaborationUser = buildNoteCollaborationUser(currentUser, t('workspace.currentUser'));
-  const canRenderBodyEditor = !shouldWaitCurrentUser;
   const findMode = useNoteFindMode({
     editorRef: bodyEditorRef,
     scopeRef: findModeScopeRef,
     canReplace: isConnected && noteInfoDisplay.canCollaborativeEdit,
   });
-  useRequest(() => interactService.recordResourceRead(resourceId), {
-    refreshDeps: [resourceId],
-  });
-
-  useRequest(() => inlineCommentSession.refresh(), {
-    pollingInterval: INLINE_COMMENT_POLLING_INTERVAL,
-    refreshDeps: [inlineCommentSession],
-  });
-
-  useUnmount(() => inlineCommentSession.destroy());
-
   const focusBody = () => {
     bodyEditorRef.current?.focus();
   };
@@ -255,70 +153,6 @@ function NoteWorkspace({ resourceId, noteInfoDisplay, onRefreshNoteInfo }: NoteW
 
   const showAiDiffDisplayModeSwitch = hasAiDiffContent;
 
-  const handlePrintPdf = useMemoizedFn(async () => {
-    const bodyApi = bodyEditorRef.current;
-    if (!bodyApi) {
-      toast.info(t('export.editorNotReady'));
-      return;
-    }
-    const titleApi = titleEditorRef.current;
-    const title = titleApi?.getPlainTitle() ?? fallbackNoteTitle ?? untitledTitle;
-    const titleRoot = titleApi?.getProseMirrorRoot() ?? null;
-    try {
-      setExportPending(true);
-      await bodyApi.exportPdf({ title, titleRoot });
-    } catch (err) {
-      toast.danger(parseErrorMessage(err));
-    } finally {
-      setExportPending(false);
-    }
-  });
-
-  const handleDownloadMarkdown = useMemoizedFn(async () => {
-    const bodyApi = bodyEditorRef.current;
-    if (!bodyApi) {
-      toast.info(t('export.editorNotReady'));
-      return;
-    }
-    try {
-      setExportPending(true);
-      const title = titleEditorRef.current?.getPlainTitle() ?? fallbackNoteTitle ?? untitledTitle;
-      const artifact = bodyApi.exportMarkdown();
-      downloadTextArtifact({
-        content: artifact.content,
-        mimeType: artifact.mimeType,
-        fileName: `${sanitizeDownloadFileName(title, untitledTitle)}.${artifact.extension}`,
-      });
-      toast.success(t('export.markdownStarted'));
-    } catch (err) {
-      toast.danger(parseErrorMessage(err));
-    } finally {
-      setExportPending(false);
-    }
-  });
-
-  const noteChatStateProvider = createNoteChatStateProvider({
-    resourceId,
-    syncStatus: status,
-    isClientContentSignaturePending: isNoteClientContentSignaturePending,
-    clientContentSignature: noteClientContentSignature,
-  });
-
-  const handleAskAi = useMemoizedFn((selection: NoteSelectionSnapshot) => {
-    setChatContext(createNoteSelectionChatContext(resourceId, selection));
-  });
-
-  const handleInlineCommentCreateRequest = useMemoizedFn((draft: NoteInlineCommentDraft) => {
-    setInlineCommentDraft(draft);
-    setResourceSidePanelMode(resourceId, 'inlineComment');
-  });
-
-  const handleInlineCommentThreadSelect = useMemoizedFn((threadId: string) => {
-    setActiveInlineCommentThreadId(threadId);
-    setInlineCommentScrollTarget({ threadId });
-    setResourceSidePanelMode(resourceId, 'inlineComment');
-  });
-
   /**
    * @wisepen-manual-effect
    * 执行时机：选中批注并完成侧栏布局更新后，将正文锚点平滑滚动到视口中央。
@@ -332,12 +166,6 @@ function NoteWorkspace({ resourceId, noteInfoDisplay, onRefreshNoteInfo }: NoteW
       threadId: inlineCommentScrollTarget.threadId,
     });
   }, [inlineCommentScrollTarget]);
-
-  const inlineCommentsBinding = {
-    session: inlineCommentSession,
-    onCreateRequest: handleInlineCommentCreateRequest,
-    onThreadSelect: handleInlineCommentThreadSelect,
-  };
 
   const resourceHostConfig = {
     className: styles.pageWrap,
